Web Service | Add articles cache and fetcher

+++ app/newsService/internal/config/config.go
@@ -36,7 +36,7 @@ type NewsManager struct {
 	FilterKeywords       []string      `yaml:"filter_keywords"`
 	FetchInterval        time.Duration `yaml:"fetch_interval"`
 	NotificationInterval time.Duration `yaml:"notification_interval"`
-	ArticlesLimit        int64         `yaml:"articles_limit"`
+	ArticlesLimit        int           `yaml:"articles_limit"`
 }
 
 func MustLoad() *Config {
